Storybook folder structure, babelrc are changed,Svgfile test deleted
