+++ spreadsheet/.babelrc
@@ -1,35 +1,22 @@
 {
-    "env": {
-        "dev": {
-            "presets": [
-                [
-                    "@babel/preset-env",
-                    {
-                        "modules": false
-                    }
-                ],
-                "@babel/preset-react"
-            ],
-            "plugins": [
-                "@babel/plugin-proposal-class-properties",
-                "@babel/plugin-transform-flow-strip-types"
-            ]
-        },
-        "test": {
-            "presets": [
-                [
-                    "@babel/preset-env",
-                    {
-                        "modules": false
-                    }
-                ],
-                "@babel/preset-react"
-            ],
-            "plugins": [
-                "@babel/plugin-proposal-class-properties",
-                "@babel/plugin-transform-modules-commonjs",
-                "@babel/plugin-transform-flow-strip-types"
-            ]
-        }
-    }
+    "presets": [
+        [
+            "@babel/preset-env",
+            {
+                "modules": false
+            }
+        ],
+        "@babel/preset-react"
+    ],
+    "plugins": [
+        "@babel/plugin-proposal-class-properties",
+        "@babel/plugin-transform-modules-commonjs",
+        "@babel/plugin-transform-flow-strip-types",
+        [
+            "@babel/plugin-transform-runtime",
+            {
+                "regenerator": true
+            }
+        ]
+    ]
 }
