--- spreadsheet/example/.eslintrc.json
@@ -1,17 +1,0 @@
-{
-  "env": {
-    "browser": true,
-    "es2020": true
-  },
-  "extends": ["eslint:recommended", "plugin:react/recommended"],
-  "parser": "babel-eslint",
-  "parserOptions": {
-    "ecmaFeatures": {
-      "jsx": true
-    },
-    "ecmaVersion": 11,
-    "sourceType": "module"
-  },
-  "plugins": ["react"],
-  "rules": {}
-}
