--- spreadsheet/example/package.json
@@ -1,31 +1,0 @@
-{
-    "name": "spreadsheet-example",
-    "homepage": ".",
-    "version": "0.0.0",
-    "private": true,
-    "scripts": {
-        "start": "node ../node_modules/react-scripts/bin/react-scripts.js start",
-        "build": "node ../node_modules/react-scripts/bin/react-scripts.js build",
-        "test": "jest -u",
-        "eject": "node ../node_modules/react-scripts/bin/react-scripts.js eject"
-    },
-    "dependencies": {
-        "react": "file:../node_modules/react",
-        "react-dom": "file:../node_modules/react-dom",
-        "react-scripts": "file:../node_modules/react-scripts",
-        "spreadsheet": "file:..",
-        "axios": "^0.19.2"
-    },
-    "devDependencies": {
-        "@babel/plugin-syntax-object-rest-spread": "^7.8.3"
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "browserslist": [
-        ">0.2%",
-        "not dead",
-        "not ie <= 11",
-        "not op_mini all"
-    ]
-}
