--- spreadsheet/example/setupTest.js
@@ -1,5 +1,0 @@
-const enzyme = require('enzyme');
-const adapter = require('enzyme-adapter-react-16');
-
-// Setup enzyme's react adapter
-enzyme.configure({ adapter: new adapter() });
