--- spreadsheet/example/src/getData.js
@@ -1,14 +1,0 @@
-import axios from "axios";
-
-export const fetchData = async (index) => {
-    const response = await axios(
-        `https://skyforce-api.azurewebsites.net/api/cargoFlightDetilsForExcel`
-    );
-    if (response && response.data && response.data.data) {
-        const { result } = response.data.data;
-        if (result) {
-            return result;
-        }
-    }
-    return [];
-};
