+++ spreadsheet/package.json
@@ -1,60 +1,22 @@
 {
     "name": "@neo-ui/spreadsheet",
-    "version": "1.0.0",
+    "version": "1.0.1",
     "description": "A library for iCargo component",
     "author": "iCargopoc",
-    "license": "MIT",
+    "license": "ISC",
     "repository": "https://github.com/iCargopoc/library",
-    "main": "dist/index.js",
-    "module": "dist/index.modern.js",
-    "source": "src/index.js",
-    "engines": {
-        "node": ">=10"
-    },
     "scripts": {
-        "build": "SET NODE_ENV=dev && microbundle-crl --no-compress --format modern,cjs",
-        "start": "SET NODE_ENV=dev && microbundle-crl watch --no-compress --format modern,cjs",
-        "flow": "flow status",
-        "storybook": "start-storybook -p 6006",
-        "build-storybook": "build-storybook",
-        "prepare": "run-s build",
         "test": "jest --coverage --env=jest-environment-jsdom-sixteen",
-        "test:cover": "jest --coverage",
-        "open:coverage": "open ./coverage/lcov-report/index.html",
         "lint": "eslint src/**/*.js src/*.js",
         "test:unit": "cross-env CI=1 react-scripts test --env=jsdom",
         "test:watch": "react-scripts test --env=jsdom",
-        "predeploy": "cd example && npm install && npm run build",
-        "deploy": "gh-pages -d example/build"
+        "storybook": "start-storybook -p 6006",
+        "build-storybook": "build-storybook",
+        "flow": "flow status"
     },
     "peerDependencies": {
-        "react": "^16.0.0",
-        "bootstrap": "^4.4.1",
-        "jspdf": "^2.0.0",
-        "jspdf-autotable": "^3.5.6",
-        "react-data-grid": "^5.0.1",
-        "react-data-grid-addons": "^6.1.0",
-        "immutability-helper": "^3.1.1",
-        "react-dnd": "^11.1.3",
-        "react-dnd-touch-backend": "^11.1.3",
-        "react-dnd-html5-backend": "^11.1.3",
-        "react-dnd-multi-backend": "^6.0.2",
-        "xlsx": "^0.16.3",
-        "node-sass": "^4.14.1",
-        "sass-loader": "^9.0.3",
-        "style-loader": "^1.2.1",
-        "react-click-away-listener": "^1.2.0",
-        "@storybook/react": "^6.0.14",
-        "@storybook/addon-docs": "^6.0.15",
-        "@panosvoudouris/addon-versions": "^1.0.1",
-        "@storybook/addon-actions": "^5.3.7",
-        "@storybook/addon-knobs": "^5.3.7",
-        "@storybook/addons": "^5.3.7",
-        "@storybook/addon-essentials": "^6.0.16",
-        "@storybook/preset-create-react-app": "^3.1.4",
-        "idempotent-babel-polyfill": "^7.4.4",
-        "lodash": "^4.17.19",
-        "prop-types": "^15.7.2"
+        "react": "^16.11.0",
+        "react-dom": "16.11.0"
     },
     "devDependencies": {
         "@babel/core": "^7.11.4",
@@ -88,18 +50,13 @@
         "eslint-plugin-react": "^7.20.5",
         "eslint-plugin-react-hooks": "^2.5.1",
         "eslint-plugin-standard": "^4.0.1",
-        "gh-pages": "^2.2.0",
         "immutability-helper": "^3.1.1",
         "jest": "^26.2.2",
         "jest-environment-jsdom-sixteen": "^1.0.3",
         "jspdf": "^2.0.0",
         "jspdf-autotable": "^3.5.6",
-        "microbundle-crl": "^0.13.10",
-        "mochawesome-report-generator": "^5.1.0",
-        "node-sass": "^4.14.1",
-        "npm-run-all": "^4.1.5",
         "prettier": "^2.0.4",
-        "react": "^16.13.1",
+        "react": "^16.11.0",
         "react-click-away-listener": "^1.2.0",
         "react-data-grid": "^5.0.1",
         "react-data-grid-addons": "^6.1.0",
@@ -107,16 +64,15 @@
         "react-dnd-html5-backend": "^11.1.3",
         "react-dnd-multi-backend": "^6.0.2",
         "react-dnd-touch-backend": "^11.1.3",
-        "react-dom": "^16.13.1",
-        "react-is": "^16.13.1",
+        "react-dom": "^16.11.0",
         "react-scripts": "^3.4.1",
-        "react-test-renderer": "^16.13.1",
-        "sass-loader": "^9.0.3",
         "style-loader": "^1.2.1",
         "xlsx": "^0.16.3",
         "idempotent-babel-polyfill": "^7.4.4",
         "lodash": "^4.17.19",
-        "prop-types": "^15.7.2"
+        "prop-types": "^15.7.2",
+        "sass": "^1.25.0",
+        "sass-loader": "^9.0.3"
     },
     "files": [
         "dist"
