+++ spreadsheet/src/Styles/grid.scss
@@ -71,7 +71,7 @@
     font-size: 14px;
     margin: 0px 0px 0px 5px;
     border: 1px solid #ddd;
-    padding: 2px 10px;
+    padding: 5px 10px;
     color: #566a81;
     cursor: pointer;
     background: #fff;
@@ -278,8 +278,8 @@ input.editor-checkbox {
     .react-grid-Toolbar {
         padding: 0;
         border: none;
-        right: 133px;
-        top: 26px;
+        right: 150px;
+        top: 42px;
         background-color: transparent;
         .btn {
             border: 1px solid #ddd;
