+++ spreadsheet/src/Utilities/SvgUtilities.js
@@ -5,140 +5,6 @@ import React from "react";
 const Svg = (p: Object): React$Element<*> => (
     <svg width="12" height="13" {...p} />
 );
-const SvgEdit = (p: Object): React$Element<*> => (
-    <svg width="12" height="13" viewBox="0 0 1792 1792" {...p} />
-);
-
-export const IconColumns = (): React$Element<*> => (
-    <Svg width="12px" height="13px">
-        <path
-            d="M.992 11.836c.045.054.1.08.162.08h4.384v-9.75H.923v9.48c0 .073.023.137.069.19zm10.016 0a.284.284 0 00.069-.19v-9.48H6.462v9.75h4.384a.207.207 0 00.162-.08zM11.661.398c.226.265.339.584.339.956v10.292c0 .372-.113.691-.339.956-.226.265-.498.398-.815.398H1.154c-.317 0-.59-.133-.815-.398A1.426 1.426 0 010 11.646V1.354C0 .982.113.663.339.398.565.133.837 0 1.154 0h9.692c.317 0 .59.133.815.398z"
-            fill="#3c476f"
-            fillOpacity="0.8"
-        />
-    </Svg>
-);
-
-export const IconAngle = (p: Object): React$Element<*> => (
-    <Svg width="12px" height="7px" {...p}>
-        <path
-            d="M11.88.722a.38.38 0 010 .553L6.277 6.88a.38.38 0 01-.554 0L.12 1.275a.38.38 0 010-.553L.721.12a.38.38 0 01.554 0L6 4.847 10.725.12a.38.38 0 01.554 0l.6.602z"
-            fill="#0e415e"
-            fillOpacity="0.4"
-        />
-    </Svg>
-);
-
-export const IconFilter = (): React$Element<*> => (
-    <Svg
-        width="11px"
-        height="11px"
-        style={{
-            position: "absolute",
-            right: "-20px",
-            top: "8px",
-            zIndex: "9"
-        }}
-    >
-        <path
-            d="M10.486 0c.218 0 .371.102.46.305.088.213.051.396-.11.547l-3.84 3.851V10.5a.467.467 0 01-.304.46.54.54 0 01-.195.04.462.462 0 01-.35-.148l-1.995-2a.481.481 0 01-.148-.352V4.703L.164.852C.003.7-.034.518.054.305A.466.466 0 01.515 0h9.972z"
-            fill="#636c8c"
-        />
-    </Svg>
-);
-
-export const IconShare = (): React$Element<*> => (
-    <Svg width="12px" height="12px">
-        <path
-            d="M7.797 7.672A2.41 2.41 0 019.5 7a2.41 2.41 0 011.77.73A2.41 2.41 0 0112 9.5a2.41 2.41 0 01-.73 1.77A2.41 2.41 0 019.5 12a2.41 2.41 0 01-1.77-.73A2.41 2.41 0 017 9.5c0-.063.005-.151.016-.266L4.203 7.828A2.41 2.41 0 012.5 8.5a2.41 2.41 0 01-1.77-.73A2.41 2.41 0 010 6c0-.693.243-1.283.73-1.77A2.41 2.41 0 012.5 3.5a2.41 2.41 0 011.703.672l2.813-1.406A3.146 3.146 0 017 2.5c0-.693.243-1.283.73-1.77A2.41 2.41 0 019.5 0a2.41 2.41 0 011.77.73A2.41 2.41 0 0112 2.5a2.41 2.41 0 01-.73 1.77A2.41 2.41 0 019.5 5a2.41 2.41 0 01-1.703-.672L4.984 5.734C4.994 5.85 5 5.938 5 6c0 .063-.005.151-.016.266l2.813 1.406z"
-            fill="#636c8c"
-        />
-    </Svg>
-);
-
-export const IconGroupSort = (): React$Element<*> => (
-    <Svg width="12px" height="12px">
-        <path
-            d="M8.013 10.346c.041.04.061.092.061.154v1.286a.207.207 0 01-.218.214H6.112a.207.207 0 01-.218-.214V10.5a.207.207 0 01.218-.214h1.744c.064 0 .116.02.157.06zm-3.271-.857c.04.04.061.091.061.154a.25.25 0 01-.068.16L2.561 11.94a.23.23 0 01-.157.06.247.247 0 01-.156-.06L.067 9.797c-.068-.072-.084-.15-.048-.235.036-.089.104-.133.204-.133h1.309V.214A.207.207 0 011.75 0h1.309a.207.207 0 01.218.214V9.43h1.308c.064 0 .116.02.157.06zm4.58-2.572c.04.04.061.092.061.154v1.286a.207.207 0 01-.218.214H6.112a.207.207 0 01-.218-.214V7.071a.207.207 0 01.218-.214h3.053c.064 0 .116.02.157.06zM10.63 3.49c.041.04.061.091.061.154v1.286a.207.207 0 01-.218.214H6.112a.207.207 0 01-.218-.214V3.643a.207.207 0 01.218-.214h4.361c.064 0 .116.02.157.06zM11.94.06c.04.04.061.092.061.154V1.5a.207.207 0 01-.218.214h-5.67a.207.207 0 01-.218-.214V.214A.207.207 0 016.112 0h5.67c.064 0 .116.02.157.06z"
-            fill="#3c476f"
-            fillOpacity="0.8"
-        />
-    </Svg>
-);
-
-export const IconSort = (p: Object): React$Element<*> => (
-    <Svg width="8px" height="5px" {...p}>
-        <path d="M0 0l4 5 4-5H0z" fill="#4f6475" />
-    </Svg>
-);
-
-export const IconEdit = (): React$Element<*> => (
-    <SvgEdit
-        height="1792"
-        viewBox="0 0 1792 1792"
-        width="1792"
-        style={{ zoom: "0.007" }}
-    >
-        <path d="M888 1184l116-116-152-152-116 116v56h96v96h56zm440-720q-16-16-33 1L945 815q-17 17-1 33t33-1l350-350q17-17 1-33zm80 594v190q0 119-84.5 203.5T1120 1536H288q-119 0-203.5-84.5T0 1248V416q0-119 84.5-203.5T288 128h832q63 0 117 25 15 7 18 23 3 17-9 29l-49 49q-14 14-32 8-23-6-45-6H288q-66 0-113 47t-47 113v832q0 66 47 113t113 47h832q66 0 113-47t47-113v-126q0-13 9-22l64-64q15-15 35-7t20 29zm-96-738l288 288-672 672H640V992zm444 132l-92 92-288-288 92-92q28-28 68-28t68 28l152 152q28 28 28 68t-28 68z" />
-    </SvgEdit>
-);
-
-export const IconPencil = (): React$Element<*> => (
-    <Svg width="10px" height="10px">
-        <path
-            d="M1.69 9.155h.706l.6-.6-1.55-1.552-.601.601v.706h.845v.845zm4.112-6.013a.153.153 0 00.046-.112c0-.097-.048-.146-.145-.146a.153.153 0 00-.112.047L2.013 6.508a.153.153 0 00-.046.112c0 .097.048.146.145.146a.153.153 0 00.112-.047l3.578-3.577zM0 7.254l5.492-5.492 2.746 2.746L2.746 10H0V7.254zm9.756-5.459a.845.845 0 01.244.601.809.809 0 01-.244.594L8.66 4.086 5.914 1.34 7.01.25A.784.784 0 017.604 0a.82.82 0 01.6.25l1.552 1.545z"
-            fill="#80a0a2"
-        />
-    </Svg>
-);
-
-export const IconTick = (): React$Element<*> => (
-    <Svg width="14px" height="11px">
-        <path
-            d="M13.747 1.519a.868.868 0 01.253.63.868.868 0 01-.253.629L7.207 9.48l-1.228 1.26a.826.826 0 01-.614.259.826.826 0 01-.614-.26L3.523 9.482.253 6.13A.868.868 0 010 5.5c0-.247.084-.457.253-.63L1.48 3.611a.826.826 0 01.614-.26c.241 0 .446.087.615.26l2.655 2.732L11.29.259A.826.826 0 0111.905 0c.24 0 .445.086.614.26l1.228 1.259z"
-            fill="#fff"
-            fillOpacity="0.996"
-        />
-    </Svg>
-);
-
-export const IconCancel = (): React$Element<*> => (
-    <Svg width="11px" height="11px">
-        <path
-            d="M10.74 8.222c.174.173.26.383.26.63a.857.857 0 01-.26.63L9.482 10.74a.857.857 0 01-.63.259.857.857 0 01-.629-.26L5.5 8.02 2.778 10.74a.857.857 0 01-.63.259.857.857 0 01-.63-.26L.26 9.482A.857.857 0 010 8.851c0-.246.086-.456.26-.629L2.98 5.5.26 2.778A.857.857 0 010 2.148c0-.247.086-.457.26-.63L1.518.26A.857.857 0 012.149 0c.246 0 .456.086.629.26L5.5 2.98 8.222.26A.857.857 0 018.852 0c.247 0 .457.086.63.26l1.259 1.259c.173.172.259.382.259.63a.857.857 0 01-.26.629L8.02 5.5l2.722 2.722z"
-            fill="gray"
-            fillOpacity="0.349"
-        />
-    </Svg>
-);
-
-export const IconSearch = (): React$Element<*> => (
-    <Svg width="11px" height="11px">
-        <path
-            d="M6.746 6.746c.58-.58.87-1.277.87-2.092 0-.815-.29-1.513-.87-2.092a2.852 2.852 0 00-2.092-.87c-.815 0-1.513.29-2.092.87-.58.58-.87 1.277-.87 2.092 0 .815.29 1.513.87 2.092.58.58 1.277.87 2.092.87.815 0 1.513-.29 2.092-.87zm4.01 2.813a.81.81 0 01.244.595c0 .229-.084.427-.251.595a.813.813 0 01-.595.251.786.786 0 01-.595-.251L7.29 8.488a4.527 4.527 0 01-2.637.82c-.63 0-1.233-.123-1.808-.367a4.653 4.653 0 01-1.488-.992 4.653 4.653 0 01-.991-1.487A4.573 4.573 0 010 4.654c0-.63.122-1.233.367-1.808.244-.575.575-1.071.991-1.488A4.653 4.653 0 012.846.367 4.573 4.573 0 014.654 0c.63 0 1.233.122 1.808.367.575.244 1.07.575 1.487.991.417.417.747.913.992 1.488.244.575.367 1.178.367 1.808 0 .97-.274 1.849-.82 2.637l2.267 2.268z"
-            fill="#3c476f"
-            fillOpacity="0.8"
-        />
-    </Svg>
-);
-
-export const RowDelete = (): React$Element<*> => (
-    <Svg width="11px" height="12px">
-        <path
-            d="M3.93 4.57c.047.047.07.107.07.18v4.5a.243.243 0 01-.07.18.243.243 0 01-.18.07h-.5a.243.243 0 01-.18-.07.243.243 0 01-.07-.18v-4.5c0-.073.023-.133.07-.18a.243.243 0 01.18-.07h.5c.073 0 .133.023.18.07zm2 0c.047.047.07.107.07.18v4.5a.243.243 0 01-.07.18.243.243 0 01-.18.07h-.5a.243.243 0 01-.18-.07.243.243 0 01-.07-.18v-4.5c0-.073.023-.133.07-.18a.243.243 0 01.18-.07h.5c.073 0 .133.023.18.07zm2 0c.047.047.07.107.07.18v4.5a.243.243 0 01-.07.18.243.243 0 01-.18.07h-.5a.243.243 0 01-.18-.07.243.243 0 01-.07-.18v-4.5c0-.073.023-.133.07-.18a.243.243 0 01.18-.07h.5c.073 0 .133.023.18.07zm1.015 6.153A.887.887 0 009 10.406V3H2v7.406a.887.887 0 00.168.528c.039.044.066.066.082.066h6.5c.016 0 .043-.022.082-.066a.7.7 0 00.113-.211zM4.133 1.086L3.75 2h3.5l-.375-.914A.22.22 0 006.742 1H4.266a.22.22 0 00-.133.086zm6.797.984c.047.047.07.107.07.18v.5a.243.243 0 01-.07.18.243.243 0 01-.18.07H10v7.406c0 .433-.122.806-.367 1.121-.245.315-.54.473-.883.473h-6.5c-.344 0-.638-.152-.883-.457C1.122 11.238 1 10.87 1 10.438V3H.25a.243.243 0 01-.18-.07.243.243 0 01-.07-.18v-.5c0-.073.023-.133.07-.18A.243.243 0 01.25 2h2.414L3.211.695c.078-.192.219-.356.422-.492C3.836.068 4.042 0 4.25 0h2.5c.208 0 .414.068.617.203.203.136.344.3.422.492L8.336 2h2.414c.073 0 .133.023.18.07z"
-            fill="#636c8c"
-        />
-    </Svg>
-);
-
-export const RowEdit = (): React$Element<*> => (
-    <Svg width="10px" height="10px">
-        <path
-            d="M1.69 9.155h.706l.6-.6-1.55-1.552-.601.601v.706h.845v.845zm4.112-6.013a.153.153 0 00.046-.112c0-.097-.048-.146-.145-.146a.153.153 0 00-.112.047L2.013 6.508a.153.153 0 00-.046.112c0 .097.048.146.145.146a.153.153 0 00.112-.047l3.578-3.577zM0 7.254l5.492-5.492 2.746 2.746L2.746 10H0V7.254zm9.756-5.459a.845.845 0 01.244.601.809.809 0 01-.244.594L8.66 4.086 5.914 1.34 7.01.25A.784.784 0 017.604 0a.82.82 0 01.6.25l1.552 1.545z"
-            fill="#636c8c"
-        />
-    </Svg>
-);
 
 export const IconClose = (): React$Element<*> => (
     <Svg width="14px" height="14px">
@@ -249,6 +115,63 @@ export const SortDelete = (): React$Element<*> => (
             d="M5.359 6.094a.32.32 0 01.096.24v6a.32.32 0 01-.096.239.336.336 0 01-.245.094h-.682a.336.336 0 01-.245-.094.32.32 0 01-.096-.24v-6a.32.32 0 01.096-.24A.336.336 0 014.432 6h.682c.1 0 .18.031.245.094zm2.727 0a.32.32 0 01.096.24v6a.32.32 0 01-.096.239.336.336 0 01-.245.094h-.682a.336.336 0 01-.245-.094.32.32 0 01-.096-.24v-6a.32.32 0 01.096-.24A.336.336 0 017.16 6h.682c.1 0 .181.031.245.094zm2.727 0a.32.32 0 01.096.24v6a.32.32 0 01-.096.239.336.336 0 01-.245.094h-.682a.336.336 0 01-.245-.094.32.32 0 01-.096-.24v-6a.32.32 0 01.096-.24A.336.336 0 019.886 6h.682c.1 0 .181.031.245.094zm1.385 8.203c.05-.129.075-.27.075-.422V4H2.727v9.875a1.16 1.16 0 00.23.703c.053.06.09.089.111.089h8.864c.021 0 .058-.03.112-.089a.928.928 0 00.154-.281zM5.636 1.447l-.522 1.22h4.772l-.511-1.22a.301.301 0 00-.181-.114H5.817a.301.301 0 00-.181.115zm9.268 1.313A.32.32 0 0115 3v.667a.32.32 0 01-.096.24.336.336 0 01-.245.093h-1.023v9.875c0 .576-.167 1.075-.5 1.495-.334.42-.735.63-1.204.63H3.068c-.469 0-.87-.203-1.204-.61-.333-.406-.5-.897-.5-1.473V4H.34a.336.336 0 01-.245-.094.32.32 0 01-.096-.24V3a.32.32 0 01.096-.24.336.336 0 01.245-.093h3.292l.746-1.74c.106-.257.298-.476.575-.656C5.23.09 5.51 0 5.795 0h3.41c.284 0 .564.09.841.27.277.181.469.4.575.657l.746 1.74h3.292c.1 0 .181.03.245.093z"
             fill="#1a4769"
             fillOpacity="0.6"
+        />
+    </Svg>
+);
+
+export const IconColumns = (): React$Element<*> => (
+    <Svg width="12px" height="13px">
+        <path
+            d="M.992 11.836c.045.054.1.08.162.08h4.384v-9.75H.923v9.48c0 .073.023.137.069.19zm10.016 0a.284.284 0 00.069-.19v-9.48H6.462v9.75h4.384a.207.207 0 00.162-.08zM11.661.398c.226.265.339.584.339.956v10.292c0 .372-.113.691-.339.956-.226.265-.498.398-.815.398H1.154c-.317 0-.59-.133-.815-.398A1.426 1.426 0 010 11.646V1.354C0 .982.113.663.339.398.565.133.837 0 1.154 0h9.692c.317 0 .59.133.815.398z"
+            fill="#3c476f"
+            fillOpacity="0.8"
+        />
+    </Svg>
+);
+
+export const IconShare = (): React$Element<*> => (
+    <Svg width="12px" height="12px">
+        <path
+            d="M7.797 7.672A2.41 2.41 0 019.5 7a2.41 2.41 0 011.77.73A2.41 2.41 0 0112 9.5a2.41 2.41 0 01-.73 1.77A2.41 2.41 0 019.5 12a2.41 2.41 0 01-1.77-.73A2.41 2.41 0 017 9.5c0-.063.005-.151.016-.266L4.203 7.828A2.41 2.41 0 012.5 8.5a2.41 2.41 0 01-1.77-.73A2.41 2.41 0 010 6c0-.693.243-1.283.73-1.77A2.41 2.41 0 012.5 3.5a2.41 2.41 0 011.703.672l2.813-1.406A3.146 3.146 0 017 2.5c0-.693.243-1.283.73-1.77A2.41 2.41 0 019.5 0a2.41 2.41 0 011.77.73A2.41 2.41 0 0112 2.5a2.41 2.41 0 01-.73 1.77A2.41 2.41 0 019.5 5a2.41 2.41 0 01-1.703-.672L4.984 5.734C4.994 5.85 5 5.938 5 6c0 .063-.005.151-.016.266l2.813 1.406z"
+            fill="#636c8c"
+        />
+    </Svg>
+);
+
+export const IconGroupSort = (): React$Element<*> => (
+    <Svg width="12px" height="12px">
+        <path
+            d="M8.013 10.346c.041.04.061.092.061.154v1.286a.207.207 0 01-.218.214H6.112a.207.207 0 01-.218-.214V10.5a.207.207 0 01.218-.214h1.744c.064 0 .116.02.157.06zm-3.271-.857c.04.04.061.091.061.154a.25.25 0 01-.068.16L2.561 11.94a.23.23 0 01-.157.06.247.247 0 01-.156-.06L.067 9.797c-.068-.072-.084-.15-.048-.235.036-.089.104-.133.204-.133h1.309V.214A.207.207 0 011.75 0h1.309a.207.207 0 01.218.214V9.43h1.308c.064 0 .116.02.157.06zm4.58-2.572c.04.04.061.092.061.154v1.286a.207.207 0 01-.218.214H6.112a.207.207 0 01-.218-.214V7.071a.207.207 0 01.218-.214h3.053c.064 0 .116.02.157.06zM10.63 3.49c.041.04.061.091.061.154v1.286a.207.207 0 01-.218.214H6.112a.207.207 0 01-.218-.214V3.643a.207.207 0 01.218-.214h4.361c.064 0 .116.02.157.06zM11.94.06c.04.04.061.092.061.154V1.5a.207.207 0 01-.218.214h-5.67a.207.207 0 01-.218-.214V.214A.207.207 0 016.112 0h5.67c.064 0 .116.02.157.06z"
+            fill="#3c476f"
+            fillOpacity="0.8"
+        />
+    </Svg>
+);
+
+export const IconSearch = (): React$Element<*> => (
+    <Svg width="11px" height="11px">
+        <path
+            d="M6.746 6.746c.58-.58.87-1.277.87-2.092 0-.815-.29-1.513-.87-2.092a2.852 2.852 0 00-2.092-.87c-.815 0-1.513.29-2.092.87-.58.58-.87 1.277-.87 2.092 0 .815.29 1.513.87 2.092.58.58 1.277.87 2.092.87.815 0 1.513-.29 2.092-.87zm4.01 2.813a.81.81 0 01.244.595c0 .229-.084.427-.251.595a.813.813 0 01-.595.251.786.786 0 01-.595-.251L7.29 8.488a4.527 4.527 0 01-2.637.82c-.63 0-1.233-.123-1.808-.367a4.653 4.653 0 01-1.488-.992 4.653 4.653 0 01-.991-1.487A4.573 4.573 0 010 4.654c0-.63.122-1.233.367-1.808.244-.575.575-1.071.991-1.488A4.653 4.653 0 012.846.367 4.573 4.573 0 014.654 0c.63 0 1.233.122 1.808.367.575.244 1.07.575 1.487.991.417.417.747.913.992 1.488.244.575.367 1.178.367 1.808 0 .97-.274 1.849-.82 2.637l2.267 2.268z"
+            fill="#3c476f"
+            fillOpacity="0.8"
+        />
+    </Svg>
+);
+
+export const IconFilter = (): React$Element<*> => (
+    <Svg
+        width="11px"
+        height="11px"
+        style={{
+            position: "absolute",
+            right: "-20px",
+            top: "8px",
+            zIndex: "9"
+        }}
+    >
+        <path
+            d="M10.486 0c.218 0 .371.102.46.305.088.213.051.396-.11.547l-3.84 3.851V10.5a.467.467 0 01-.304.46.54.54 0 01-.195.04.462.462 0 01-.35-.148l-1.995-2a.481.481 0 01-.148-.352V4.703L.164.852C.003.7-.034.518.054.305A.466.466 0 01.515 0h9.972z"
+            fill="#636c8c"
         />
     </Svg>
 );
